requirements.txt
numpy
fastapi
mangum
uvicorn
tensorflow
transformers
tf-keras
